chore: tracer refactor code structure
