+++ pyproject.toml
@@ -61,6 +61,7 @@ dev = [
     "tavily-python>=0.7.5",
     "langgraph>=0.4.3",
     "pre-commit>=4.2.0",
+    "types-requests>=2.32.4.20250611",
 ]
 
 [tool.hatch.build]
