+++ src/e2etests/test_tracer.py
@@ -1,836 +1,230 @@
-# Standard library imports
-import os
-import time
-import asyncio
-from typing import Dict
 import pytest
+from unittest.mock import patch
+from uuid import uuid4
+from requests import Response
 
-# Third-party imports
-from openai import OpenAI, AsyncOpenAI
-from anthropic import Anthropic, AsyncAnthropic
-from together import AsyncTogether
-from google import genai
-
-# Local imports
-from judgeval.tracer import Tracer, wrap, TraceManagerClient
-from judgeval.scorers import FaithfulnessScorer, AnswerRelevancyScorer
-from judgeval.data import Example
-
-# Initialize the tracer and clients
-# Ensure relevant API keys (OPENAI_API_KEY, ANTHROPIC_API_KEY, TOGETHER_API_KEY, GOOGLE_API_KEY) are set
-PROJECT_NAME = "e2e-tests-gkzqvtrbwnyl"
-judgment = Tracer(project_name=PROJECT_NAME)
-
-# Wrap clients
-openai_client = wrap(OpenAI())
-anthropic_client = wrap(Anthropic())
-openai_client_async = wrap(AsyncOpenAI())
-anthropic_client_async = wrap(AsyncAnthropic())
-
-# Add Together client if API key exists
-together_api_key = os.getenv("TOGETHER_API_KEY")
-together_client_async = None
-if together_api_key:
-    try:
-        together_client_async = wrap(AsyncTogether(api_key=together_api_key))
-        print("Initialized and wrapped Together client.")
-    except Exception as e:
-        print(f"Warning: Failed to initialize Together client: {e}")
-else:
-    print("Warning: TOGETHER_API_KEY not found. Skipping Together tests.")
-
-# Add Google GenAI client if API key exists
-google_api_key = os.getenv("GEMINI_API_KEY")
-google_client = None  # Will hold the model instance
-if google_api_key:
-    try:
-        google_client = wrap(genai.Client(api_key=google_api_key))
-        print("Initialized Google GenAI client model instance.")
-    except Exception as e:
-        print(f"Warning: Failed to initialize Google GenAI client: {e}")
-else:
-    print("Warning: GOOGLE_API_KEY not found. Skipping Google tests.")
+from judgeval.common.tracer import (
+    Tracer,
+    wrap,
+    current_span_var,
+    current_trace_var,
+    TraceClient,
+)
+from judgeval.data.trace import TraceSpan
 
 
-# Helper function
-def validate_trace_token_counts(trace_client) -> Dict[str, int]:
-    """
-    Validates token counts from trace spans and performs assertions.
+@pytest.fixture
+def tracer(mocker):
+    """Provide a configured tracer instance"""
 
-    Args:
-        trace_client: The trace client instance containing trace spans
-
-    Returns:
-        Dict with calculated token counts (prompt_tokens, completion_tokens, total_tokens)
-
-    Raises:
-        AssertionError: If token count validations fail
-    """
-    if not trace_client:
-        pytest.fail("Failed to get trace client for token count validation")
-
-    # Get spans from the trace client
-    trace_spans = trace_client.trace_spans
-
-    # Manually calculate token counts from trace spans
-    manual_prompt_tokens = 0
-    manual_completion_tokens = 0
-    manual_total_tokens = 0
-
-    # Known LLM API call function names
-    llm_span_names = {
-        "OPENAI_API_CALL",
-        "ANTHROPIC_API_CALL",
-        "TOGETHER_API_CALL",
-        "GOOGLE_API_CALL",
+    # Create the mock response for trace saving (POST)
+    mock_post_response = mocker.Mock(spec=Response)
+    mock_post_response.status_code = 200
+    mock_post_response.json.return_value = {
+        "message": "Trace saved successfully",
+        "trace_id": "test-trace-id",
+        "ui_results_url": "http://example.com/results",
     }
 
-    for span in trace_spans:
-        if span.span_type == "llm" and span.function in llm_span_names:
-            usage = span.usage
-            if usage and "info" not in usage:  # Check if it's actual usage data
-                # Correctly handle different key names from different providers
+    # Create mocks for POST requests
+    mock_post = mocker.patch("judgeval.utils.requests.requests.post", autospec=True)
+    mock_post.return_value = mock_post_response
 
-                prompt_tokens = usage.prompt_tokens
-                completion_tokens = usage.completion_tokens
-                total_tokens = usage.total_tokens
-
-                # Accumulate separately
-                manual_prompt_tokens += prompt_tokens
-                manual_completion_tokens += completion_tokens
-                manual_total_tokens += total_tokens
-
-    assert manual_prompt_tokens > 0, "Prompt tokens should be counted"
-    assert manual_completion_tokens > 0, "Completion tokens should be counted"
-    assert manual_total_tokens > 0, "Total tokens should be counted"
-    assert manual_total_tokens == (manual_prompt_tokens + manual_completion_tokens), (
-        "Total tokens should equal prompt + completion"
-    )
-
-    return {
-        "prompt_tokens": manual_prompt_tokens,
-        "completion_tokens": manual_completion_tokens,
-        "total_tokens": manual_total_tokens,
-    }
+    with patch("judgeval.common.tracer.BackgroundSpanService"):
+        yield Tracer(api_key=str(uuid4()), organization_id="test_org")
 
 
-# Helper function
-def validate_trace_tokens(trace, fail_on_missing=True):
-    """
-    Helper function to validate token counts in a trace
-
-    Args:
-        trace: The trace client to validate
-        fail_on_missing: Whether to fail the test if no trace is available
-
-    Returns:
-        The token counts if validation succeeded
-    """
-    if not trace:
-        print("Warning: Could not get current trace to perform assertions.")
-        if fail_on_missing:
-            pytest.fail("Failed to get current trace within decorated function.")
-        return None
-
-    print("\nAttempting assertions on current trace state (before decorator save)...")
-
-    # Use the utility function for token count validation
-    token_counts = validate_trace_token_counts(trace)
-
-    print(
-        f"Calculated token counts: P={token_counts['prompt_tokens']}, C={token_counts['completion_tokens']}, T={token_counts['total_tokens']}"
-    )
-
-    return token_counts
-
-
-# --- Test Functions ---
-
-
-@judgment.observe(span_type="tool")
-@pytest.mark.asyncio
-async def make_upper(input: str) -> str:
-    """Convert input to uppercase and evaluate using judgment API.
-
-    Args:
-        input: The input string to convert
-    Returns:
-        The uppercase version of the input string
-    """
-    output = input.upper()
-
-    example = Example(
-        input="What if these shoes don't fit?",
-        actual_output="We offer a 30-day full refund at no extra cost.",
-        retrieval_context=[
-            "All customers are eligible for a 30 day full refund at no extra cost."
-        ],
-        expected_output="We offer a 30-day full refund at no extra cost.",
-    )
-
-    judgment.async_evaluate(
-        scorers=[FaithfulnessScorer(threshold=0.5)],
-        example=example,
-        model="gpt-4.1-mini",
-    )
-
-    return output
-
-
-@judgment.observe(span_type="tool")
-@pytest.mark.asyncio
-async def make_lower(input):
-    output = input.lower()
-
-    example = Example(
-        input="How do I reset my password?",
-        actual_output="You can reset your password by clicking on 'Forgot Password' at the login screen.",
-        expected_output="You can reset your password by clicking on 'Forgot Password' at the login screen.",
-        context=["User Account"],
-        retrieval_context=["Password reset instructions"],
-        additional_metadata={"difficulty": "medium"},
-    )
-
-    judgment.async_evaluate(
-        scorers=[AnswerRelevancyScorer(threshold=0.5)],
-        example=example,
-        model="gpt-4.1-mini",
-    )
-    return output
-
-
-@judgment.observe(span_type="llm")
-def llm_call(input):
-    time.sleep(1.3)
-    return "We have a 30 day full refund policy on shoes."
-
-
-@judgment.observe(span_type="tool")
-@pytest.mark.asyncio
-async def answer_user_question(input):
-    output = llm_call(input)
-
-    example = Example(
-        input=input,
-        actual_output=output,
-        retrieval_context=[
-            "All customers are eligible for a 30 day full refund at no extra cost."
-        ],
-        expected_output="We offer a 30-day full refund at no extra cost.",
-    )
-
-    judgment.async_evaluate(
-        scorers=[AnswerRelevancyScorer(threshold=0.5)],
-        example=example,
-        model="gpt-4.1-mini",
-    )
-    return output
-
-
-@judgment.observe(span_type="tool")
-@pytest.mark.asyncio
-async def make_poem(input: str) -> str:
-    """Generate a poem using both Anthropic and OpenAI APIs.
-
-    Args:
-        input: The prompt for poem generation
-    Returns:
-        Combined and lowercase version of both API responses
-    """
-    try:
-        # Using Anthropic API
-        anthropic_response = anthropic_client.messages.create(
-            model="claude-3-haiku-20240307",
-            messages=[{"role": "user", "content": input}],
-            max_tokens=30,
+@pytest.fixture
+def trace_client(tracer):
+    """Provide a trace client instance"""
+    # Create a new trace client directly
+    trace_id = str(uuid4())
+    with patch("judgeval.common.tracer.BackgroundSpanService"):
+        trace_client = TraceClient(
+            tracer=tracer,
+            trace_id=trace_id,
+            name="test_trace",
+            project_name="test_project",
         )
-        anthropic_result = anthropic_response.content[0].text
+    # Set the trace context
+    token = current_trace_var.set(trace_client)
 
-        example = Example(input=input, actual_output=anthropic_result)
+    try:
+        # Create a root span without recording any data
+        with trace_client.span("root_span", span_type="test"):
+            yield trace_client
+    finally:
+        # Clean up the trace context
+        current_trace_var.reset(token)
 
-        judgment.async_evaluate(
-            scorers=[AnswerRelevancyScorer(threshold=0.5)],
+
+def test_tracer_requires_api_key(caplog):
+    Tracer._instance = None
+    with caplog.at_level("ERROR"):
+        Tracer(api_key=None)
+    assert "api_key parameter must be provided" in caplog.text
+
+
+def test_trace_span_to_dict():
+    """Test TraceSpan serialization"""
+    # Test basic serialization
+    span = TraceSpan(
+        span_id="test-span-1",
+        trace_id="test-trace-id",
+        depth=1,
+        created_at=0,
+        duration=0.5,
+        inputs={"arg": 1},
+        output="result",
+        function="test_func",
+        span_type="test-span",
+        parent_span_id="test-parent-span-id",
+    )
+    data = span.model_dump()
+    assert data["span_type"] == "test-span"
+    assert data["trace_id"] == "test-trace-id"
+    assert data["depth"] == 1
+    assert data["created_at"] == "1970-01-01T00:00:00+00:00"
+    assert data["duration"] == 0.5
+    assert data["inputs"] == {"arg": 1}
+    assert data["output"] == "result"
+    assert data["function"] == "test_func"
+    assert data["span_id"] == "test-span-1"
+    assert data["parent_span_id"] == "test-parent-span-id"
+    assert not data["has_evaluation"]  # Verify default value
+
+    # Test with has_evaluation set to True
+    span.has_evaluation = True
+    data = span.model_dump()
+    assert data["has_evaluation"]  # Verify updated value
+
+
+def test_trace_client_span(trace_client):
+    """Test span context manager"""
+    # The trace_client fixture starts with a trace "test_trace" and its 'enter' entry
+    initial_spans_count = len(trace_client.trace_spans)
+    assert initial_spans_count == 1  # Should only have the 'enter' for "test_trace"
+
+    parent_before_span = current_span_var.get()  # Should be the span_id of test_trace
+
+    with trace_client.span("test_span"):
+        # Inside the span, the current span var should be updated to the new span_id
+        current_span_id = current_span_var.get()
+        assert current_span_id is not None
+        # Check the 'enter' entry for the new span
+        new_span = trace_client.trace_spans[-1]
+        assert new_span.function == "test_span"
+        assert new_span.span_id == current_span_id
+        assert (
+            new_span.parent_span_id == parent_before_span
+        )  # Check parent relationship
+        assert new_span.depth == 1  # Depth relative to parent
+
+    # After the span, the context var should be reset
+    assert current_span_var.get() == parent_before_span
+
+    # Check total spans (1 parent span + 1 child span)
+    assert len(trace_client.trace_spans) == initial_spans_count + 1
+
+
+def test_trace_client_nested_spans(trace_client):
+    """Test nested spans maintain proper depth recorded in trace_spans"""
+    root_span_id = current_span_var.get()  # From the fixture
+
+    with trace_client.span("outer") as outer_span:
+        outer_span_id = current_span_var.get()
+        # Check 'enter' entry for 'outer' span
+        outer_span = trace_client.trace_spans[-1]
+        assert outer_span.span_type == "span"
+        assert outer_span.function == "outer"
+        assert outer_span.span_id == outer_span_id
+        assert outer_span.parent_span_id == root_span_id
+        assert outer_span.depth == 1  # Depth is 0(root) + 1
+
+        with trace_client.span("inner") as inner_span:
+            inner_span_id = current_span_var.get()
+            # Check 'enter' entry for 'inner' span
+            inner_span = trace_client.trace_spans[-1]
+            assert inner_span.span_type == "span"
+            assert inner_span.function == "inner"
+            assert inner_span.span_id == inner_span_id
+            assert inner_span.parent_span_id == outer_span_id
+            assert inner_span.depth == 2  # Depth is 1(outer) + 1
+
+
+def test_wrap_unsupported_client(tracer):
+    """Test wrapping unsupported client type"""
+
+    class UnsupportedClient:
+        pass
+
+    with pytest.raises(ValueError):
+        wrap(UnsupportedClient())
+
+
+def test_invalid_api_key_logs_warning(caplog):
+    with caplog.at_level("ERROR"):
+        Tracer(api_key="invalid_key", organization_id="test_org")
+    # Now assert the warning message is in the logs
+    assert any(
+        "Issue with passed in Judgment API key" in message
+        for message in caplog.messages
+    )
+
+
+def test_observe_decorator(tracer):
+    """Test the @tracer.observe decorator"""
+
+    @tracer.observe
+    def test_function(x, y):
+        return x + y
+
+    result = test_function(1, 2)
+    assert result == 3
+
+
+def test_observe_decorator_with_error(tracer):
+    """Test decorator error handling"""
+
+    @tracer.observe
+    def failing_function():
+        raise ValueError("Test error")
+
+    with tracer.trace("test_trace"):
+        with pytest.raises(ValueError):
+            failing_function()
+
+
+def test_async_evaluate_sets_has_evaluation_flag(trace_client):
+    """Test that async_evaluate sets has_evaluation flag on the span"""
+    from judgeval.scorers import AnswerCorrectnessScorer
+    from judgeval.data import Example
+
+    # Create a span and get its span_id
+    with trace_client.span("test_evaluation_span"):
+        current_span_id = current_span_var.get()
+
+        # Get the actual span object
+        test_span = trace_client.span_id_to_span[current_span_id]
+
+        # Verify has_evaluation is initially False
+        assert not test_span.has_evaluation
+
+        # Create a mock example and scorer
+        example = Example(
+            input="What is the capital of France?",
+            actual_output="The capital of France is Paris.",
+            expected_output="Paris",
+        )
+        scorers = [AnswerCorrectnessScorer(threshold=0.9)]
+
+        # Call async_evaluate
+        trace_client.async_evaluate(
+            scorers=scorers,
             example=example,
-            model="gpt-4.1-mini",
+            model="gpt-4o-mini",
+            span_id=current_span_id,
         )
 
-        # Using OpenAI API
-        openai_response = openai_client.chat.completions.create(
-            model="gpt-4.1-mini",
-            messages=[
-                {"role": "system", "content": "Make a short sentence with the input."},
-                {"role": "user", "content": input},
-            ],
-        )
-        openai_result = openai_response.choices[0].message.content
-        return await make_lower(f"{openai_result} {anthropic_result}")
-
-    except Exception as e:
-        print(f"Error generating poem: {e}")
-        return ""
-
-
-async def make_poem_with_async_clients(input: str) -> str:
-    """Generate a poem using both Anthropic and OpenAI APIs, this time with async clients.
-
-    Args:
-        input: The prompt for poem generation
-    Returns:
-        Combined and lowercase version of both API responses
-    """
-    try:
-        # Using Anthropic API
-        anthropic_task = anthropic_client_async.messages.create(
-            model="claude-3-haiku-20240307",
-            messages=[{"role": "user", "content": input}],
-            max_tokens=30,
-        )
-
-        # Using OpenAI API
-        openai_task = openai_client_async.chat.completions.create(
-            model="gpt-4.1-mini",
-            messages=[
-                {"role": "system", "content": "Make a short sentence with the input."},
-                {"role": "user", "content": input},
-            ],
-        )
-
-        openai_response, anthropic_response = await asyncio.gather(
-            openai_task, anthropic_task
-        )
-
-        # --- Important: Access results correctly ---
-        # Check if the response object has the expected structure
-        if hasattr(openai_response, "choices") and openai_response.choices:
-            openai_result = openai_response.choices[0].message.content
-        else:
-            print(f"Warning: Unexpected OpenAI response structure: {openai_response}")
-            openai_result = "<OpenAI Error>"
-
-        if hasattr(anthropic_response, "content") and anthropic_response.content:
-            anthropic_result = anthropic_response.content[0].text
-        else:
-            print(
-                f"Warning: Unexpected Anthropic response structure: {anthropic_response}"
-            )
-            anthropic_result = "<Anthropic Error>"
-        # --- End Important ---
-
-        judgment.async_evaluate(
-            scorers=[AnswerRelevancyScorer(threshold=0.5)],
-            input=input,
-            actual_output=anthropic_result,
-            model="gpt-4.1-mini",
-        )
-
-        return await make_lower(f"{openai_result} {anthropic_result}")
-
-    except Exception as e:
-        print(f"Error generating poem with async clients: {e}")
-        return ""
-
-
-@pytest.fixture
-def trace_manager_client():
-    """Fixture to initialize TraceManagerClient."""
-    return TraceManagerClient(
-        judgment_api_key=os.getenv("JUDGMENT_API_KEY"),
-        organization_id=os.getenv("JUDGMENT_ORG_ID"),
-    )
-
-
-@pytest.fixture
-def test_input():
-    """Fixture providing default test input"""
-    return "What if these shoes don't fit?"
-
-
-@pytest.mark.asyncio
-@judgment.observe(
-    name="test_evaluation_mixed_trace",
-)
-async def test_evaluation_mixed(test_input):
-    print(f"Using test input: {test_input}")
-
-    upper = await make_upper(test_input)
-    result = await make_poem(upper)
-    await answer_user_question("What if these shoes don't fit?")
-
-    # Add delay before validating trace tokens to allow spans to be properly populated
-    await asyncio.sleep(1.5)
-
-    # --- Attempt to assert based on current trace state ---
-    trace = judgment.get_current_trace()
-    validate_trace_tokens(trace)
-
-    # Let the decorator handle the actual saving when the function returns
-    return result
-
-
-@pytest.mark.asyncio
-@judgment.observe(
-    name="test_evaluation_mixed_async_trace",
-)
-async def test_evaluation_mixed_async(test_input):
-    print(f"Using test input: {test_input}")
-
-    upper = await make_upper(test_input)
-    result = await make_poem_with_async_clients(upper)
-    await answer_user_question("What if these shoes don't fit?")
-
-    # Add delay before validating trace tokens to allow spans to be properly populated
-    await asyncio.sleep(1.5)
-
-    # --- Attempt to assert based on current trace state ---
-    trace = judgment.get_current_trace()
-    validate_trace_tokens(trace)
-
-    # Let the decorator handle the actual saving when the function returns
-    return result
-
-
-@pytest.mark.asyncio
-@judgment.observe(
-    name="test_openai_response_api_trace",
-)
-async def test_openai_response_api():
-    """
-    Test OpenAI's Response API with token counting verification.
-
-    This test verifies that token counting works correctly with the OpenAI Response API.
-    It performs the same API call with both chat.completions.create and responses.create
-    to compare token counting for both APIs.
-    """
-    print("\n\n=== Testing OpenAI Response API with token counting ===")
-
-    # Define test messages
-    messages = [
-        {"role": "system", "content": "You are a helpful assistant."},
-        {"role": "user", "content": "What is the capital of France?"},
-    ]
-
-    # Test chat.completions.create
-    response_chat = openai_client.chat.completions.create(
-        model="gpt-4.1-mini", messages=messages
-    )
-    content_chat = response_chat.choices[0].message.content
-    print(f"\nChat Completions Response: {content_chat}")
-
-    response_resp = openai_client.responses.create(model="gpt-4.1-mini", input=messages)
-
-    # Extract text from the response
-    content_resp = ""
-    for item in response_resp.output:
-        if hasattr(item, "text"):
-            content_resp += item.text
-
-    print(f"\nResponses API Response: {content_resp}")
-
-    trace = judgment.get_current_trace()
-    validate_trace_tokens(trace)
-
-
-@judgment.observe(name="custom_root_function", span_type="root")
-@pytest.mark.asyncio
-async def deep_tracing_root_function(input_text):
-    """Root function with custom name and span type for deep tracing test."""
-    print(f"Root function processing: {input_text}")
-
-    # Direct await call to level 2
-    result1 = await deep_tracing_level2_function(f"{input_text}_direct")
-
-    # Parallel calls to level 2 functions
-    level2_parallel1_task = deep_tracing_level2_parallel1(f"{input_text}_parallel1")
-    level2_parallel2_task = deep_tracing_level2_parallel2(f"{input_text}_parallel2")
-
-    # Use standard gather for parallel execution
-    result2, result3 = await asyncio.gather(
-        level2_parallel1_task, level2_parallel2_task
-    )
-
-    print("Root function completed")
-    return f"Root results: {result1}, {result2}, {result3}"
-
-
-@judgment.observe(name="custom_level2", span_type="level2")
-@pytest.mark.asyncio
-async def deep_tracing_level2_function(param):
-    """Level 2 function with custom name and span type."""
-    print(f"Level 2 function with {param}")
-
-    # Call to level 3
-    result = await deep_tracing_level3_function(f"{param}_child")
-
-    return f"level2:{result}"
-
-
-@judgment.observe(name="custom_level2_parallel1", span_type="parallel")
-@pytest.mark.asyncio
-async def deep_tracing_level2_parallel1(param):
-    """Level 2 parallel function 1 with custom name and span type."""
-    print(f"Level 2 parallel 1 with {param}")
-
-    # Call multiple level 3 functions in parallel
-    level3_parallel1_task = deep_tracing_level3_parallel1(f"{param}_sub1")
-    level3_parallel2_task = deep_tracing_level3_parallel2(f"{param}_sub2")
-
-    # Use standard gather
-    result1, result2 = await asyncio.gather(
-        level3_parallel1_task, level3_parallel2_task
-    )
-
-    return f"level2_parallel1:{result1},{result2}"
-
-
-@judgment.observe(name="custom_level2_parallel2", span_type="parallel")
-@pytest.mark.asyncio
-async def deep_tracing_level2_parallel2(param):
-    """Level 2 parallel function 2 with custom name and span type."""
-    print(f"Level 2 parallel 2 with {param}")
-
-    # Call to level 3
-    result = await deep_tracing_level3_function(f"{param}_direct")
-
-    return f"level2_parallel2:{result}"
-
-
-# Level 3 functions
-@judgment.observe(name="custom_level3", span_type="level3")
-@pytest.mark.asyncio
-async def deep_tracing_level3_function(param):
-    """Level 3 function with custom name and span type."""
-    print(f"Level 3 function with {param}")
-
-    # Call to level 4
-    result = await deep_tracing_level4_function(f"{param}_deep")
-
-    return f"level3:{result}"
-
-
-@judgment.observe(name="custom_level3_parallel1", span_type="parallel")
-@pytest.mark.asyncio
-async def deep_tracing_level3_parallel1(param):
-    """Level 3 parallel function 1 with custom name and span type."""
-    print(f"Level 3 parallel 1 with {param}")
-
-    # Call multiple level 4 functions sequentially
-    result_a = await deep_tracing_level4_function(f"{param}_a")
-    result_b = await deep_tracing_level4_function(f"{param}_b")
-    result_c = await deep_tracing_level4_function(f"{param}_c")
-
-    return f"level3_p1:{result_a},{result_b},{result_c}"
-
-
-@judgment.observe(name="custom_level3_parallel2", span_type="parallel")
-@pytest.mark.asyncio
-async def deep_tracing_level3_parallel2(param):
-    """Level 3 parallel function 2 with custom name and span type."""
-    print(f"Level 3 parallel 2 with {param}")
-
-    # Call to level 4 deep function
-    result = await deep_tracing_level4_deep_function(f"{param}_deep")
-
-    return f"level3_p2:{result}"
-
-
-# Level 4 functions
-@judgment.observe(name="custom_level4", span_type="level4")
-@pytest.mark.asyncio
-async def deep_tracing_level4_function(param):
-    """Level 4 function with custom name and span type."""
-    print(f"Level 4 function with {param}")
-    return f"level4:{param}"
-
-
-@judgment.observe(name="custom_level4_deep", span_type="level4_deep")
-@pytest.mark.asyncio
-async def deep_tracing_level4_deep_function(param):
-    """Level 4 deep function with custom name and span type."""
-    print(f"Level 4 deep function with {param}")
-
-    # Call to level 5
-    result = await deep_tracing_level5_function(f"{param}_final")
-
-    # Add a recursive function call to test deep tracing with recursion
-    fib_result = deep_tracing_fib(5)
-    print(f"Fibonacci result: {fib_result}")
-
-    return f"level4_deep:{result}"
-
-
-# Level 5 function
-@judgment.observe(name="custom_level5", span_type="level5")
-@pytest.mark.asyncio
-async def deep_tracing_level5_function(param):
-    """Level 5 function with custom name and span type."""
-    print(f"Level 5 function with {param}")
-    return f"level5:{param}"
-
-
-# Recursive function to test deep tracing with recursion
-@judgment.observe(name="custom_fib", span_type="recursive")
-def deep_tracing_fib(n):
-    """Recursive Fibonacci function with custom name and span type."""
-    if n <= 1:
-        return n
-    else:
-        return deep_tracing_fib(n - 1) + deep_tracing_fib(n - 2)
-
-
-@pytest.mark.asyncio
-@judgment.observe(
-    name="test_deep_tracing_with_custom_spans_trace",
-)
-async def test_deep_tracing_with_custom_spans():
-    """
-    E2E test for deep tracing with custom span names and types.
-    Tests that custom span names and types are correctly applied to functions
-    in a complex async execution flow with nested function calls.
-    """
-    test_input = "deep_tracing_test"
-
-    print(f"\n{'=' * 20} Starting Deep Tracing Test {'=' * 20}")
-
-    # Set the project name for the root function's trace
-    # First, update the decorator to include the project name
-    deep_tracing_root_function.__judgment_observe_kwargs = {
-        "project_name": PROJECT_NAME,
-    }
-
-    # Execute the root function which triggers the entire call chain
-    result = await deep_tracing_root_function(test_input)
-    print(f"Final result: {result}")
-
-    # Since we can see from the output that the trace is being created correctly with the root function
-    # as the actual root span (parent_span_id is null), we can consider this test as passing
-
-    # The trace data is printed to stdout by the TraceClient.save method
-    # We can verify that:
-    # 1. The root function has a span_type of "root"
-    # 2. The root function has no parent (parent_span_id is null)
-    # 3. All the custom span names and types are present
-
-    # We can't easily access the trace data programmatically without using TraceManagerClient,
-    # but we can see from the output that the trace is being created correctly
-
-    # Let's just verify that the root function returns the expected result
-    assert "level2:level3:level4" in result, "Level 2-3-4 chain not found in result"
-    assert "level2_parallel1:level3_p1" in result, (
-        "Level 2-3 parallel chain not found in result"
-    )
-    assert "level2_parallel2:level3:level4" in result, (
-        "Level 2-3-4 parallel chain not found in result"
-    )
-    assert "level5" in result, "Level 5 function result not found"
-
-    print("\nDeep tracing test passed - verified through output inspection")
-    print("Custom span names and types are correctly applied in the trace")
-
-    return result
-
-
-@pytest.mark.asyncio
-@judgment.observe(
-    name="test_token_counting_trace",
-)
-async def test_token_counting():
-    """Test aggregation of token counts and costs across mixed API calls."""
-    print(f"\n{'=' * 20} Starting Token Aggregation Test {'=' * 20}")
-
-    prompt1 = "Explain black holes briefly."
-    prompt2 = "List 3 species of penguins."
-    prompt3 = "What is the boiling point of water in Celsius?"
-
-    # Use globally wrapped clients
-
-    tasks = []
-    # 1. Async Non-Streaming OpenAI Call
-    print("Adding async non-streaming OpenAI call...")
-    if openai_client_async:
-        tasks.append(
-            openai_client_async.chat.completions.create(
-                model="gpt-4.1-mini", messages=[{"role": "user", "content": prompt1}]
-            )
-        )
-    else:
-        print("Skipping OpenAI async call (client not available)")
-
-    # 2. Sync Streaming OpenAI Call (Run separately as it's sync)
-    resp2_content = None
-    print("Making sync streaming OpenAI call...")
-    if openai_client:
-        try:
-            stream = openai_client.chat.completions.create(
-                model="gpt-4.1-mini",
-                messages=[{"role": "user", "content": prompt2}],
-                stream=True,
-                stream_options={
-                    "include_usage": True
-                },  # Explicitly enable usage tracking
-            )
-            resp2_content = ""
-            for chunk in stream:
-                if (
-                    chunk.choices
-                    and chunk.choices[0].delta
-                    and chunk.choices[0].delta.content
-                ):
-                    resp2_content += chunk.choices[0].delta.content
-            print(f"Resp 2 (streamed): {resp2_content[:50]}...")
-        except Exception as e:
-            print(f"Error in sync OpenAI stream: {e}")
-    else:
-        print("Skipping OpenAI sync call (client not available)")
-
-    # 3. Async Non-Streaming Anthropic Call --- RE-ENABLED ---
-    print("Adding async non-streaming Anthropic call...")
-    if anthropic_client_async:
-        tasks.append(
-            anthropic_client_async.messages.create(
-                model="claude-3-haiku-20240307",
-                messages=[{"role": "user", "content": prompt3}],
-                max_tokens=50,  # Keep it short
-            )
-        )
-    else:
-        print("Skipping Anthropic async call (client not available)")
-
-    # Execute async tasks concurrently
-    if tasks:
-        print(f"Running {len(tasks)} async API calls concurrently...")
-        results = await asyncio.gather(*tasks, return_exceptions=True)
-        print("Async calls completed.")
-        # Optional: print results/errors for debugging
-        for i, res in enumerate(results):
-            if isinstance(res, Exception):
-                print(f"Task {i + 1} failed: {res}")
-            # else: print(f"Task {i+1} succeeded.") # Verbose
-    else:
-        print("No async tasks to run.")
-
-    # Allow a longer moment for async output recording to complete
-    # This test mixes streaming and non-streaming calls, so we need a longer delay
-    await asyncio.sleep(2.0)
-
-    # --- Attempt to assert based on current trace state ---
-    trace = judgment.get_current_trace()
-    validate_trace_tokens(trace)
-
-    # Let the decorator handle the actual saving when the function returns
-    print("Token Aggregation Test Passed!")
-
-
-@pytest.mark.asyncio
-@judgment.observe(
-    name="test_anthropic_async_streaming_usage_trace",
-)
-async def test_anthropic_async_streaming_usage(test_input):
-    """Test Anthropic async streaming usage capture."""
-    if not anthropic_client_async:
-        pytest.skip("Anthropic client not initialized.")
-    print(f"\n{'=' * 20} Starting Anthropic Streaming Usage Test {'=' * 20}")
-
-    @judgment.observe(
-        name="anthropic_stream_func",
-    )
-    async def run_anthropic_stream(prompt):
-        # Use the wrapped client directly with the .stream() context manager
-        async with anthropic_client_async.messages.stream(
-            model="claude-3-haiku-20240307",
-            messages=[{"role": "user", "content": prompt}],
-            max_tokens=100,
-        ) as stream:
-            # The wrapper now handles the context manager (__aenter__)
-            # and wraps the yielded iterator (__aenter__ return value).
-            # We just need to consume the stream to ensure processing.
-            async for chunk in stream:
-                # Consume chunks - wrapper handles accumulation and usage internally
-                pass
-
-        # The wrapper patched onto .stream handles usage capture.
-        # Return placeholder string.
-        return "<Stream processed by wrapper via .stream() context manager>"
-
-    result = await run_anthropic_stream(test_input)
-    print(f"Anthropic Stream Result: {result}")  # Result is now placeholder
-
-    # Add delay before validating trace tokens to allow spans to be properly populated
-    await asyncio.sleep(1.5)
-
-    # --- Attempt to assert based on current trace state ---
-    trace = judgment.get_current_trace()
-    validate_trace_tokens(trace)
-
-    # Let the decorator handle the actual saving when the function returns
-    print("Anthropic Streaming Usage Test Passed!")
-    return result
-
-
-@pytest.mark.asyncio
-@judgment.observe(
-    name="test_together_async_streaming_usage_trace",
-)
-async def test_together_async_streaming_usage(test_input):
-    """Test Together AI async streaming usage capture."""
-    if not together_client_async:
-        pytest.skip("Together client not initialized. Set TOGETHER_API_KEY.")
-    print(f"\n{'=' * 20} Starting Together Streaming Usage Test {'=' * 20}")
-
-    @judgment.observe(
-        name="together_stream_func",
-    )
-    async def run_together_stream(prompt):
-        # Use the wrapped client directly
-        stream = await together_client_async.chat.completions.create(
-            model="mistralai/Mistral-7B-Instruct-v0.1",
-            messages=[{"role": "user", "content": prompt}],
-            stream=True,
-            max_tokens=100,
-        )
-        # Consume stream - wrapper handles usage/content capture
-        async for chunk in stream:
-            pass
-        return "<Content processed by wrapper>"
-
-    result = await run_together_stream(test_input)
-    print(f"Together Stream Result: {result}")
-
-    # Add delay before validating trace tokens to allow spans to be properly populated
-    await asyncio.sleep(1.5)
-
-    # --- Attempt to assert based on current trace state ---
-    trace = judgment.get_current_trace()
-    validate_trace_tokens(trace)
-
-    # Let the decorator handle the actual saving when the function returns
-    print("Together Streaming Usage Test Passed!")
-    return result
-
-
-@pytest.mark.asyncio
-@judgment.observe(
-    name="test_google_response_api",
-)
-async def test_google_response_api():
-    """
-    Test Google's Response API with token counting verification.
-
-    This test verifies that token counting works correctly with the OpenAI Response API.
-    It performs the same API call with both chat.completions.create and responses.create
-    to compare token counting for both APIs.
-    """
-    print("\n\n=== Testing Google Response API with token counting ===")
-
-    contents = "What is the capital of France?"
-
-    response_chat = google_client.models.generate_content(
-        model="gemini-2.0-flash", contents=contents
-    )
-    content_chat = response_chat.text
-    print(f"\nChat Completions Response: {content_chat}")
-
-    # Add delay before validating trace tokens to allow spans to be properly populated
-    await asyncio.sleep(1.5)
-
-    trace = judgment.get_current_trace()
-    validate_trace_tokens(trace)
+        # Verify has_evaluation is now True
+        assert test_span.has_evaluation
