+++ src/judgeval/common/storage/__init__.py
@@ -1,0 +1,6 @@
+from judgeval.common.storage.s3_storage import S3Storage
+
+
+__all__ = [
+    "S3Storage",
+]
